refactor(server): split cloud sync event stores
