+++ scripts/max_lines_allowlist.txt
@@ -29,7 +29,6 @@ server/proliferate/auth/identity/service.py 807 existing auth service exceeds re
 server/proliferate/db/store/automations.py 816 server structure migration debt for oversized store file
 server/proliferate/db/store/billing.py 2605 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_sync/commands.py 1565 server structure migration debt for oversized store file
-server/proliferate/db/store/cloud_sync/events.py 1109 server structure migration debt for oversized store file
 server/proliferate/db/store/cloud_workspaces.py 1187 server structure migration debt for oversized store file
 server/proliferate/server/billing/service.py 2459 server structure migration debt for oversized service file during billing orchestration consolidation
 server/tests/integration/schema_migration_assertions.py 878 existing server oversized test helper
